# voice-customer-service-lab/packages/contracts/openapi/voice-api.v1.json
{
  "components": {
    "schemas": {
...
      "def-12": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "enum": [
                  "mock"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "volcengine"
                ],
                "type": "string"
              }
            ]
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "room_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "rtc_user_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "session_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/def-11"
          }
        },
        "required": [
          "session_id",
          "room_id",
          "rtc_user_id",
          "provider",
          "state",
          "revision",
          "created_at",
          "expires_at"
        ],
        "title": "SessionSnapshot",
        "type": "object"
      },
      "def-13": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "app_id": {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              "expires_at": {
                "format": "date-time",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "mock"
                ],
                "type": "string"
              },
              "room_id": {
                "maxLength": 80,
                "minLength": 8,
                "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
                "type": "string"
              },
              "token": {
                "minLength": 8,
                "type": "string"
              },
              "user_id": {
                "maxLength": 80,
                "minLength": 8,
                "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
                "type": "string"
              }
            },
            "required": [
              "kind",
              "app_id",
              "room_id",
              "user_id",
              "token",
              "expires_at"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "app_id": {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              "expires_at": {
                "format": "date-time",
                "type": "string"
              },
              "kind": {
                "enum": [
                  "volcengine"
                ],
                "type": "string"
              },
              "room_id": {
                "maxLength": 80,
                "minLength": 8,
                "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
                "type": "string"
              },
              "token": {
                "minLength": 8,
                "type": "string"
              },
              "user_id": {
                "maxLength": 80,
                "minLength": 8,
                "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
                "type": "string"
              }
            },
            "required": [
              "kind",
              "app_id",
              "room_id",
              "user_id",
              "token",
              "expires_at"
            ],
            "type": "object"
          }
        ],
        "title": "RtcCredentials"
      },
...
      "def-15": {
        "additionalProperties": false,
        "properties": {
          "locale": {
            "enum": [
              "zh-CN"
            ],
            "type": "string"
          }
        },
        "required": [
          "locale"
        ],
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "def-16": {
        "additionalProperties": false,
        "properties": {
          "command_replayed": {
            "type": "boolean"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/def-14"
            },
            "type": "array"
          },
          "rtc_credentials": {
            "$ref": "#/components/schemas/def-13"
          },
          "session": {
            "$ref": "#/components/schemas/def-12"
          }
        },
        "required": [
          "session",
          "rtc_credentials",
          "events",
          "command_replayed"
        ],
        "title": "CreateSessionResponse",
        "type": "object"
      },
      "def-17": {
        "additionalProperties": false,
        "properties": {
          "command_replayed": {
            "type": "boolean"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/def-14"
            },
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/def-12"
          },
          "summary": {
            "$ref": "#/components/schemas/def-32"
          }
        },
        "required": [
          "session",
          "events",
          "summary",
          "command_replayed"
        ],
        "title": "EndSessionResponse",
        "type": "object"
      },
...
      "def-21": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "MockTurnRequest",
        "type": "object"
      },
      "def-22": {
        "anyOf": [
          {
            "enum": [
              "user_request"
            ],
            "type": "string"
          },
          {
            "enum": [
              "unsupported_request"
            ],
            "type": "string"
          },
          {
            "enum": [
              "safety_concern"
            ],
            "type": "string"
          },
          {
            "enum": [
              "repeated_failure"
            ],
            "type": "string"
          }
        ],
        "title": "HandoffReason"
      },
      "def-23": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/def-22"
          }
        },
        "required": [
          "reason"
        ],
        "title": "HandoffRequest",
        "type": "object"
      },
      "def-24": {
        "additionalProperties": false,
        "properties": {
          "command_replayed": {
            "type": "boolean"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/def-14"
            },
            "type": "array"
          },
          "handoff": {
            "$ref": "#/components/schemas/def-25"
          },
          "session": {
            "$ref": "#/components/schemas/def-12"
          },
          "summary": {
            "$ref": "#/components/schemas/def-32"
          }
        },
        "required": [
          "session",
          "events",
          "summary",
          "handoff",
          "command_replayed"
        ],
        "title": "HandoffResponse",
        "type": "object"
      },
...
      "def-31": {
        "additionalProperties": false,
        "properties": {
          "command_replayed": {
            "type": "boolean"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/def-14"
            },
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/def-12"
          }
        },
        "required": [
          "session",
          "events",
          "command_replayed"
        ],
        "title": "SessionCommandResponse",
        "type": "object"
      },
      "def-32": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "enum": [
                  "completed"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "handoff_requested"
                ],
                "type": "string"
              }
            ]
          },
          "raw_audio_retained": {
            "enum": [
              false
            ],
            "type": "boolean"
          },
          "retention_expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "sensitive_input_detected": {
            "type": "boolean"
          },
          "session_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "summary_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/def-33"
            },
            "maxItems": 3,
            "type": "array",
            "uniqueItems": true
          },
          "transcript_retained": {
            "enum": [
              false
            ],
            "type": "boolean"
          },
          "turn_count": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "summary_id",
          "session_id",
          "outcome",
          "turn_count",
          "topics",
          "sensitive_input_detected",
          "raw_audio_retained",
          "transcript_retained",
          "generated_at",
          "retention_expires_at"
        ],
        "title": "SessionPrivacySummary",
        "type": "object"
      },
      "def-33": {
        "anyOf": [
          {
            "enum": [
              "general_support"
            ],
            "type": "string"
          },
          {
            "enum": [
              "order_status"
            ],
            "type": "string"
          },
          {
            "enum": [
              "human_handoff"
            ],
            "type": "string"
          }
        ],
        "title": "SessionSummaryTopic"
      },
...
      "def-6": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/def-5"
          },
          "command_replayed": {
            "type": "boolean"
          }
        },
        "required": [
          "agent",
          "command_replayed"
        ],
        "title": "AgentCommandResponse",
        "type": "object"
      },
      "def-7": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]*$",
                "type": "string"
              },
              "correlation_id": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "correlation_id"
            ],
            "title": "ApiErrorDetail",
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Voice Customer Service API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/sessions": {
      "post": {
        "operationId": "createVoiceSession",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-15"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-16"
                }
              }
            },
            "description": "Default Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-16"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}": {
      "delete": {
        "operationId": "endVoiceSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-17"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/agent": {
      "delete": {
        "operationId": "stopVoiceAgent",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-6"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "startVoiceAgent",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-6"
                }
              }
            },
            "description": "Default Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-6"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/handoff": {
      "post": {
        "operationId": "requestHumanHandoff",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-23"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-24"
                }
              }
            },
            "description": "Default Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-24"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "session-closure"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/mock-turns": {
      "post": {
        "operationId": "submitMockVoiceTurn",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-21"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-31"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "mock-sessions"
        ]
      }
    },
...
  }
}